Update make_meta.sh
Took out processes that are now in the make_dirs.sh script.

Changed the case statements to record a variable based on the user's choice, instead of running the command immediately after the user input. 

Added if statements to read the assigned variable and run the commands after all of the case statements.