Update make_meta.sh
Certain functions were still running, even when the "no" case had been selected. This was resolved by changing the if statements before the individual tool case statements to from "if [[ $var=0]]" to "if [[ "$var" -eq "0" ]]. I'm not sure if it was the -eq or the "" around the var and the zero. Similarly the if statements at the end of the script that call the functions needed "" around the var and the 1, or they wouldn't work.