Update HMSG_auto.config
Changed the ArtFile variable path in the MakeArtworkFile function to end in the "ArtistLast Name, ArtistFirstName" directory, instead of the "accession_title" directory to match the path assigned if the Art File is selected from the result of the find command. Previously the MakeArtworkFile function would assign the variable to a directory that was lower in the directory structure hierarchy than this, while the find command could not define this variable.