Update make_dirs.sh
removed comment prompting me to change the ArtFile variable in the if loop beginning on line 20. I had initially thought that I would change this ArtFile path to match the path that had been in the MakeArtworkFile function, but that requires having the accession number and title. Instead, I have changed the MakeArtworkFunction to define the ArtFile variable in the same way that this if loop does in the "else" option. See recent commits to the config file for details