Update make_dirs.sh
A user dragged and dropped a volume into terminal to input the name, and the trailing space on the end of the input was interpreted later in the script as a "\" which broke some of the CLI tools called in make_meta.sh. I've added a sed command after the "read" volume input to clear any spaces from the input.