Update HMSG_auto.config
Updated log commands in each function (with the exception of the few functions that don't write to the log) to write logs in hierarchy trees to be more human readable. Each function also has a "results" section written to the log that summarizes the results once function is completed (essentially success or "no sidecars found").